techpack: display: add a correction to dimlayer interpolation
Our dimlayer is calculated using a linear model, regardless a delay of 10 was added to the panel on both to avoid brightness flashing (panel is slow to react), on FOD-HBM ON/OFF. This causes the model to be unaccurate to generate the alpha value for dimlayer. Add .01 as equivalent correction factor for this delay with a +- .0025 error. This fixes the black lut at FOD Pressed/Unpressed.

Signed-off-by: Carlos Ayrton Lopez Arroyo <15030201@itcelaya.edu.mx>